+++ endpoint/onramp/controllers/src/main/java/com/hotels/road/onramp/OnrampSwaggerConfiguration.java
@@ -24,6 +24,7 @@ import java.util.List;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.PathSelectors;
@@ -42,7 +43,8 @@ public class OnrampSwaggerConfiguration {
   private static final String BASIC_AUTH = "basicAuth";
 
   @Bean
-  SecurityConfiguration security() {
+  @Primary
+  SecurityConfiguration onrampSecurity() {
     return SecurityConfigurationBuilder
         .builder()
         .realm("Data Highway")