+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Switched all metrics to micrometer API with prometheus actuator (except truck park).
+- Fixed tests working around the disabling of bean overriding in Spring Boot 2.1.x
 
 ## [6.0.2] - 2019-01-23
 ### Changed